Hand patch shader files to fix remaining compile errors.
Some remaining errors that happen when the game compiles these HLSL shaders.  The errors make the shader be skipped, so hand patching them makes them compile and run, and gives a complete set of shaders for the game.

Using these shaders in ShaderFixes folder will make AC3 load and run with no compiler or decompiler errors in the log.

These are the list of remaining Decompiler bugs that generate bad code for the AC3 shaders.