Fix Decompiler bug with icb r4.x.xyzw format
The icb entry for the CBuffer was a different format than all other named entries, and was missing the [0] that would make the new fix work.

Changed 'icb' to 'icb[0]' so that my new replacement code will find the proper subscript in the icb case.   It's a special case, created in ParseDefinitions as a -1 index.