Add game specific d3dx.ini as play target.
The d3dx.ini has many settings, some conflict with the goal of having it run fast.  This allows the AC3 version to be different than debug or release version, and makes it less likely to botch an AC3 release version.

This file is tuned for maximum performance, including no shader hunting.