Fix icb[] errors, and bitmask with bfi instruction error.
icb was defined wrong, and apparently never used before, although the code is there.  changed to
const float4 icb[]
to make it compile same code.

bfi instruction was decoded like documentation said, but the fxc compiler doesn't handle uint properly. Weak.  Changed bitmask format to a different style so that it gets around the uint bug.

This fixes 6 shaders, and fixes bitmasks in a few others.