Add a cache of shaders from SR3, from master branch code.
These shaders were all produced from the master branch code, before all the changes for fixing warnings.
The game is working mostly smoothly in master, so these make a good unit-test for the shaders being created by this warnings branch.

If a shader changes from branch to branch, that's nearly certain to be an introduced bug in the warnings branch.