Fix two glitches in Alien shaders, white-skey, dark-shader.
White sky seemed to be an output conflict, fixed by Mike's sharing a cb1 buffer directly with PS, so we can comment out the VS output and still work. Leaving it here in project with comments, rather than removing it.

Setting rasterizer_disable_scissor=0 to fix the dark shader problem.  Not sure why, but it fixes problem at the arrival scene in Sevastopol.