More solid setup for menus and HUD in Mordor.
These are all shaders used through the UI, now set up to handle aimMode and playMode.  When in playMode ,the depth is user selectable for their HUD, using the F4 key.

When in aimMode, we override just the aiming shaders to max depth.

Update d3dx.ini for key handling as well.

Trying to get out of assembly mindset, and using some language here to make it more clear.