Add VS for dialog choices round rects.
This VS is modified to force the element to screen depth, but only when the x2=1, which will be set from the d3dx.ini texture filtering.  

The PS is not used here, because it will clip the texture if moved, but it is the one specified in the texture filtering.