Two more uint returns from Load
Convert the variables returned to uint4 temporary variables.  They must be uint, because that is how the original buffer or texture is defined.  

Also updates all references, until they are redefined.  In particular the u_r7 usage was carefully matched to use case, where there are some that are uint, and some that are float, intermixed.