Last loop cleanup
This is fairly good.  The code generation is very close to the original, with no obvious problems.  Lots of register names change as the compiler switches them around, but all major instructions are correct.

Unfortunately, it still does not work correctly, with half the screen black.  And no further obvious problems with the code.