Remove faulty (int) coercions
The instruction like:
  i_r13.xy = (int)r4.zw;
is wrong, because that's a 2 element int coercion, specified as one.  This generates bad code, as it truncates to just the z parameter.  

The coercion is generated by the Decompiler to match the conversions in the code, but after changing the destination type to an Int4, there is no need for it, as it's clear to fxc that it needs to convert.  

Fixing this problem clears the half-screen black problem.  The broken part here was using the depth map in x only, without y.